RealtimeSTT-master/requirement.txt
numpy
pandas
streamlit 
openai
python-dotenv
streamlit-folium
streamlit-geolocation
fastapi==0.115.0
langchain
langchain-openai
langchain-community==0.3.17
unstructured
python-magic-bin
faiss-cpu
pyttsx3
whisper
sounddevice